Added Deutsche Telekom Root CA 2.
Deutsche Telekom Root CA 2 is soon to be included in Mozilla's trust store
and it is already included in Microsoft's.  It is owned by T-Telesec,
a subsidiary of Deutsche Telekom.  This CA certifies other CAs, like
e.g. the DFN PKI, which certifies state and federal institutions like
universities in Germany.

Signed-off-by: Philipp Kern <pkern@debian.org>